requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
twilio
httpx
google-auth
google-auth-oauthlib
google-api-python-client
python-dateutil
tzdata